# perf: Remove the quadratic behavior in haskellPackages.ghcWithPackages