@sternenseemann 2022-10-04
`shellFor` has some extra code to merge multiple `ghcWithPackages` environment, maybe that gets tripped up somehow.