@FRidh 2022-10-04
Could this function be put in lib so we can convert other users as well?