@guibou 2022-10-04
@FRidh It could yes. In which `lib` subfile do you think it will be at home?