@FRidh 2022-10-04
If the behaviour is the same it could replace the original function. Or is there any reason for not doing that?