@guibou 2022-10-04
Will debugging, the new function is missing some values:

I'm trying to understand. That's weird because it is able to build my packages, so there may be a weird corner case I did not anticpiated.

@FRidh I'm not sure the new function is as general as the old one. Because the old was able to return anything which is a `set`, but the new can only return things which are `set` with an `outPath`.

```diff
1d0
< /nix/store/014dicp6v8l44jwx0qiadmrk23j4xyk3-hspec-core-2.8.5
3,5d1
< /nix/store/0hgxgp65g0mg5mw58hdikhj0zzwwx7ni-hspec-discover-2.8.5
< /nix/store/1nv8b47dm0z8xf55jfzpmzsrz9fhkyjs-ansi-terminal-0.11.3
< /nix/store/1v197gcfjm47549wadipcbkbv2cmbiv3-profunctors-5.6.2
8,10d3
< /nix/store/3rmmdpg12m6ryrjdpwspyrcbhpwjzq51-StateVar-1.2.2
< /nix/store/4q8w5bsn9phpkz3wbhhxpci1lgjbhjf6-tf-random-0.5
< /nix/store/4ydclrfs2kxbq9195n7hw1qm8nbmnssy-hspec-expectations-0.8.2
12,16d4
< /nix/store/64xbi7926z9sv977p3gv421dnfb50xhi-strict-0.4.0.1
< /nix/store/6dx6pwy5pqx08b72filzbb3b0k5zlyj3-call-stack-0.4.0
< /nix/store/6i2fvi1akhfbm006dm3fgvf9xg2p9cw8-contravariant-1.5.5
< /nix/store/6qmqark6jxbglzbjwmyvn8l25a8ibgzm-quickcheck-io-0.2.0
< /nix/store/6vicys7f43ikc0ych1hmaln7702bbrxy-bifunctors-5.5.13
18d5
< /nix/store/846xg53r9fivzkdmpmr5yh6vkfy196f1-transformers-base-0.4.6
20,23d6
< /nix/store/8rwafhh7nrzifzb2m58rm7naxkarpza5-extensible-exceptions-0.1.1.4
< /nix/store/8vbs4n20qbh95m22sixy279k52cyzcj5-these-1.1.1.1
< /nix/store/99sr0597428hb7ib0b13ijw243agkngn-scientific-0.3.7.0
< /nix/store/bhwzkqg404gnlj70cjgjqcphbqxwfs0p-colour-2.3.6
26,27d8
< /nix/store/cq2i4wjw061sgn2qhzcgf3hky6mw991b-setenv-0.1.1.3
< /nix/store/dfv4ai9hi2f9avlammpf1qyhhslhmjxb-assoc-1.0.2
29,30d9
< /nix/store/fh0dc4ndn60b475hb1l36kpym5i7yhkk-regex-tdfa-1.3.2
< /nix/store/fnh9rwax3x0r8qjdh6jiv5yyy44ignbg-splitmix-0.1.0.4
33,34d11
< /nix/store/gdarb3wg3y2wi824wbnl2hvsc3v7np65-clock-0.8.3
< /nix/store/h9z3amqgmrpdlrc8bmcms6mf3nqy8dhj-old-locale-1.0.0.7
37d13
< /nix/store/hzg99hggqw6g48zlfwwqa6mrkhp6jawf-OneTuple-0.3.1
39,40d14
< /nix/store/i4rylsmkbkgzvffkvpbgdppacbf8wz7v-QuickCheck-2.14.2
< /nix/store/ipq5z3xa1d84ica0qfn15z96r9hv9xmh-xml-1.3.14
42d15
< /nix/store/jsxhrq9yjfq8s84xdrdad8zhrhzkqpqn-th-abstraction-0.4.5.0
44,45d16
< /nix/store/kgi26v8anani7ja7svrqj94dn1hgyri0-kan-extensions-5.2.5
< /nix/store/kh71s1yn68wib6kzm8m3r24v222ymabj-integer-logarithms-1.0.3.1
47,48d17
< /nix/store/l89yqkcjssmjwiggj84znl2ac8mkaadv-regex-posix-0.96.0.1
< /nix/store/lq8ilf068imv54np2l1s4zh91n1vkgv6-parallel-3.2.2.0
50d18
< /nix/store/mvfvj06vpiyjv4g25i4p490kjnfr4jpg-free-5.1.9
54d21
< /nix/store/p8cr9aav175i6lqicjffjjva58zkcyi0-regex-base-0.94.0.2
56d22
< /nix/store/w4bia9by4srqb0zz9zn01d64hawnhrd3-invariant-0.5.6
58,63d23
< /nix/store/wasm9bckfrxm9qgwfng0xvpc3nlikfxi-primitive-0.7.3.0
< /nix/store/wqxdrq2bi7wzjh561gl94k2bnik0v68z-indexed-traversable-instances-0.1.1.1
< /nix/store/yhpsk4bxql6dfnc8m86wd29y237in2zm-binary-orphans-1.0.3
< /nix/store/yp5fr91qkgh0vh3bvkgfgzkzrxiksq7g-ansi-wl-pprint-0.6.9
< /nix/store/z1xnays4jk1c09b70728rw3hpnwzsjqq-hostname-1.0
< /nix/store/z653l9xrs9hnv16ymiaa3naj5q6xcjqv-pretty-terminal-0.1.0.0
65d24
< /nix/store/zgqkpq7j5yrvnyjh09988i2iy42mahbp-comonad-5.0.8
```