@guibou 2022-10-04
Haha, I'm just stupid. Simple mistake when copy pasting (I've refactored the `item.propagatedNativeBuildInputs or [ ]`  and missed the `item.val`).