@guibou 2022-10-04
@ofborg build tests.haskell.shellFor