@guibou 2022-10-04
The tests should now pass (at least they pass locally).

As suggested by @FRidh, I replaced the `lib.closePropagation` by the new implementation (with a conditional based on the existence of `builtins.genericClosure`).

Looks like the behavior should be highly similar, except for the ordering of the results. There are not much usage of this function:

```
nixos/modules/services/audio/mopidy.nix:    paths = closePropagation cfg.extensionPackages;
pkgs/applications/editors/eclipse/default.nix:            filter (x: x ? isEclipsePlugin) (closePropagation plugins);
pkgs/development/compilers/yosys/default.nix:      paths = lib.closePropagation plugins;
pkgs/development/haskell-modules/hoogle.nix:  docPackages = lib.closePropagation
pkgs/development/haskell-modules/with-packages-wrapper.nix:  paths         = lib.filter (x: x ? isHaskellLibrary) (lib.closePropagation packages);
pkgs/development/idris-modules/with-packages.nix:let paths = lib.closePropagation packages;
pkgs/development/tools/misc/hydra/unstable.nix:    paths = with perlPackages; lib.closePropagation
pkgs/top-level/perl-packages.nix:  makeFullPerlPath = deps: makePerlPath (lib.misc.closePropagation deps);
```

A quick look makes me think that we may be safe to do this change on the complete nixpkgs. However it may trigger a massive rebuild, so perhaps it should be merged to another branch. What do you think?