@sternenseemann 2022-10-05
@guibou Please add a changelog entry for this to the release notes for 21.11, since it is a breaking change (which is probably not too worrysome, since probably very few downstream users will be using it without having an `outPath` attribute). It should mention:

1. Improved performance of `lib.closePropagation` for e.g. `ghcWithPackages`.
2. The breakage, i.e. that every set needs an `outPath` now.