@guibou 2022-10-05
@sternenseemann documentation added (note, I changed the notes for 22.11 instead of 21.11 as you suggested. Please tell me if you were really meaning 21.11). Thank you @teto who also told me to regenerate the db thing.